fix: pull up logic for resolving deep values into base class
