+++ packages/core/src/entity/EditorModeEntityStore.ts
@@ -95,78 +95,6 @@ export class EditorModeEntityStore extends EditorEntityStore {
     if (!entityLink || !entityLink.sys) return;
 
     const fieldValue = super.getValue(entityLink, path);
-
-    // walk around to render asset files
-    return fieldValue && typeof fieldValue == 'object' && (fieldValue as AssetFile).url
-      ? (fieldValue as AssetFile).url
-      : fieldValue;
-  }
-
-  public getValueDeep(
-    headLinkOrEntity: UnresolvedLink<'Entry' | 'Asset'>,
-    deepPath: string
-  ): string | undefined {
-    const resolveFieldset = (
-      unresolvedFieldset: Array<[null, string, string?]>,
-      headEntity: Entry | Asset
-    ) => {
-      const resolvedFieldset: Array<[Entry | Asset, string, string?]> = [];
-      let entityToResolveFieldsFrom: Entry | Asset = headEntity;
-      for (let i = 0; i < unresolvedFieldset.length; i++) {
-        const isLeaf = i === unresolvedFieldset.length - 1; // with last row, we are not expecting a link, but a value
-        const row = unresolvedFieldset[i];
-        const [_, field, _localeQualifier] = row;
-        if (!entityToResolveFieldsFrom) {
-          throw new Error(
-            `Logic Error: Cannot resolve field ${field} of a fieldset as there is no entity to resolve it from.`
-          );
-        }
-
-        if (isLeaf) {
-          resolvedFieldset.push([entityToResolveFieldsFrom, field, _localeQualifier]);
-          return resolvedFieldset;
-        }
-
-        const fieldValue = get<string>(entityToResolveFieldsFrom, ['fields', field]) as
-          | UnresolvedLink<'Entry' | 'Asset'>
-          | undefined
-          | any;
-
-        if (undefined === fieldValue) {
-          throw new Error(`Cannot resolve field ${field} of a fieldset as it is not defined.`);
-        } else if (isLink(fieldValue)) {
-          let entity: Asset | Entry | undefined = this.getEntityFromLink(fieldValue);
-          if (entity === undefined) {
-            throw new Error(
-              `Logic Error: Cannot resolve field ${field} of a fieldset row [${JSON.stringify(
-                row
-              )}] as the link is broken. ${JSON.stringify({ link: fieldValue })}`
-            );
-          }
-
-          resolvedFieldset.push([entityToResolveFieldsFrom, field, _localeQualifier]);
-          entityToResolveFieldsFrom = entity; // we move up
-        } else {
-          throw new Error(
-            `LogicError: Invalid value of a field we consider a reference field. Cannot resolve field ${field} of a fieldset as it is not a link, neither undefined.`
-          );
-        }
-      }
-      return resolvedFieldset;
-    };
-
-    const headEntity: Entry | Asset | undefined = isLink(headLinkOrEntity)
-      ? this.getEntityFromLink(headLinkOrEntity)
-      : headLinkOrEntity;
-    if (undefined === headEntity) {
-      return;
-    }
-    const unresolvedFieldset = parseDataSourcePathIntoFieldset(deepPath);
-    const resolvedFieldset = resolveFieldset(unresolvedFieldset, headEntity);
-
-    // To get values take last, leaf entity and field
-    const [leafEntity, field, _localeQualifier] = resolvedFieldset[resolvedFieldset.length - 1];
-    const fieldValue = get<string>(leafEntity, ['fields', field]); // is allowed to be undefined (when non-required field not set; or even when field does NOT exist on the type)
     return transformAssetFileToUrl(fieldValue);
   }
 }
